fix: buttons in mobile view
