+++ features/welcome/early-adoption-banner/styles.tsx
@@ -1,4 +1,4 @@
-import { Block } from '@lidofinance/lido-ui';
+import { Block, Button } from '@lidofinance/lido-ui';
 import { Stack } from 'shared/components';
 import styled from 'styled-components';
 
@@ -26,8 +26,15 @@ export const BlockStyled = styled(Block)`
       rgba(249, 129, 183, 0) 100%
     ),
     linear-gradient(97deg, #00a3ff 36.36%, #2238ff 99.58%);
+`;
 
-  a {
-    flex: 1 0 48%;
+export const StyledButton = styled(Button)`
+  min-width: max-content;
+
+  span {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 6px; // @style
   }
 `;
