+++ features/welcome/welcome.tsx
@@ -7,6 +7,7 @@ import { useCsmPublicRelease } from 'shared/hooks/useCsmStatus';
 import { Connect, Fallback } from 'shared/wallet';
 import { EarlyAdoptionBanner } from './early-adoption-banner';
 import { WelcomeSection } from './welcome-section';
+import styled from 'styled-components';
 
 export const Welcome: FC = () => {
   const { active, isConnected } = useAccount();
@@ -19,22 +20,26 @@ export const Welcome: FC = () => {
       {isWrongChain && <Fallback />}
       <WelcomeSection>
         <Stack wrap>
-          <Connect
+          <ConnectStyle
             fullwidth
             matomoEvent={MATOMO_CLICK_EVENTS_TYPES.connectAsNodeOperator}
           >
             I am a Node Operator
-          </Connect>
-          <Connect
+          </ConnectStyle>
+          <ConnectStyle
             fullwidth
             matomoEvent={MATOMO_CLICK_EVENTS_TYPES.connectToBecomeNodeOperator}
             color="secondary"
           >
             Become a Node Operator
-          </Connect>
+          </ConnectStyle>
         </Stack>
       </WelcomeSection>
       {!isPublicRelease && <EarlyAdoptionBanner />}
     </>
   );
 };
+
+const ConnectStyle = styled(Connect)`
+  min-width: max-content;
+`;
