--- shared/components/link-button/style.ts
@@ -1,11 +1,0 @@
-import { Button } from '@lidofinance/lido-ui';
-import styled from 'styled-components';
-
-export const StyledButton = styled(Button)`
-  span {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 6px; // @style
-  }
-`;
